task15

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,21 +9,21 @@
 
 int main(void)
 {
-	int num1 = 0;
-	int num2 = 1;
-	int sum = 0;
+	long int num1 = 0;
+	long int num2 = 1;
+	long int sum = 0;
 	int n = 0;
 
 	while (n < 49)
 	{
 		sum = num1 + num2;
-		printf("%d, ", sum);
+		printf("%li, ", sum);
 		num1 = num2;
 		num2 = sum;
 
 		n++;
 	}
 	sum = num1 + num2;
-	printf("%d\n", sum);
+	printf("%li\n", sum);
 	return (0);
 }
